Adds basic authentication flow
Implements a basic authentication flow with a login screen
and a profile view. Uses an `AuthManager` to manage the user's
login state.

Introduces `AuthManager` as an `ObservableObject` to handle
authentication logic and persist user session data.

Adapts the `ContentView` to display either the `LoginView` or
`ProfileView` based on the `isLoggedIn` state of the
`AuthManager`.

Updates to use `NavigationStack` on iOS 16 and above.